requirements.txt
Django>=4.1.1,<4.2.0
djangorestframework>=3.13.1,<3.14.0
psycopg2-binary>=2.9.3,<2.10.0
python-dotenv==0.21.0

